requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework-api-key==2.*
djangorestframework==3.12.4
dj_static==0.0.6
psycopg2==2.8.6
whitenoise==5.2.0
gunicorn==20.0.4
requests==2.25.1
django-environ==0.4.5
